Added Styling Changes for Responsive Design
