+++ src/components/Footer.js
@@ -1,10 +1,12 @@
-import React, { useEffect } from "react";
+import React from "react";
 import BottomNavigation from '@mui/material/BottomNavigation';
 import BottomNavigationAction from '@mui/material/BottomNavigationAction';
 import TvIcon from "@mui/icons-material/Tv";
 import MovieIcon from "@mui/icons-material/Movie";
 import ArrowCircleUpIcon from '@mui/icons-material/ArrowCircleUp';
 import { useHistory } from "react-router-dom";
+
+import '../styles/Footer.css';
 
 const styles = {
   root: {
@@ -17,26 +19,11 @@ const styles = {
 };
 
 export default function Footer() {
-  const [value, setValue] = React.useState(0);
   const history = useHistory();
 
-  // useEffect(() => {
-  //   if (value === 0) {
-  //     history.push("/");
-  //   } else if (value === 1) {
-  //     history.push("/tv-series");
-  //   } else if (value === 2) {
-  //     window.scroll(0, 0);
-  //   }
-  // }, [value, history]);
-
   return (
-      <footer>
+      <footer className='footer'>
            <BottomNavigation
-                value={value}
-                // onChange={(event, newValue) => {
-                //     setValue(newValue);
-                // }}
                 showLabels
                 style={styles.root} >
                 <BottomNavigationAction
