+++ src/routes.js
@@ -1,6 +1,5 @@
 import Home from "./pages/Home";
 import MovieDetail from "./pages/MovieDetail";
-import MovieForm from "./pages/MovieForm";
 
 export const routes = [
   {
@@ -13,12 +12,6 @@ export const routes = [
     path: "/movies/:id",
     exact: false,
     component: <MovieDetail />,
-    title: "MovieDetail"
-  },
-  {
-    path: "/movie/form",
-    exact: false,
-    component: <MovieForm />,
     title: "MovieDetail"
   }
 ];
