+++ src/styles/Footer.css
@@ -1,0 +1,3 @@
+footer.footer {
+    /* flex: 0 0 auto; */
+}
