+++ src/styles/Layout.css
@@ -1,18 +1,22 @@
 .container {
-  width: 1200px;
-  margin: 30px auto 0;
+  width: 75rem;
+  margin: 1.9rem auto 0;
 }
 
 * {
   scroll-behavior: smooth;
 }
 
-body.layout-body {
-  min-height: 60vh;
+body {
   background-color: #39445a;
+}
+
+body.layout-body {
+  /* min-height: 60vh;
+  background-color: #39445a; */
   color: white;
-  padding-top: 130px;
-  padding-bottom: 70px;
+  padding-top: 8.5rem;
+  padding-bottom: 4.4rem;
 }
 
 @media (max-width: 700px) {
