+++ src/styles/MovieDetail.css
@@ -1,5 +1,5 @@
 .detail-container {
-    margin-top: 40px;
+    margin-top: 2.5rem;
     padding: 10px;
     display: flex;
     flex-direction: row;
@@ -7,36 +7,56 @@
     box-shadow: 0px 6px 18px rgba(0, 0, 0, 0.1);
     border-radius: 10px;
     color:rgba(255, 255, 255, 0.6);
-  }
+}
+
+.poster img {
+  border-radius: 10px;
+}
+
+.info {
+  width: 100%;
+}
+
+.field {
+  margin-left: 2.5rem;
+  text-align: left;
+  font-size: 22px;
+  margin-bottom: 1.25rem;
+}
+
+.field:first-child {
+  margin-bottom: 0.2rem;
+}
+
+.label {
+  color:rgba(255, 255, 255, 0.8);
+  font-size: 20px;
+  font-weight: bold;
+}
+
+.label-p{
+  display:inline-block;
+  color:rgba(255, 255, 255, 0.6);
+  font-size: 20px;
+}
   
-  .poster img {
-    border-radius: 10px;
+.title{
+  color:white;
+  font-size: 28px;
+}
+
+.circular-loader {
+  margin-top: 3rem;
+  text-align: center;
+}
+
+@media screen and (max-width: 800px) {
+  .detail-container {
+      flex-direction: column;
+      align-items: center;
   }
-  
-  .info {
-    width: 100%;
-  }
-  
+
   .field {
-    margin-left: 40px;
-    text-align: left;
-    font-size: 22px;
-    margin-bottom: 20px;
+    margin-left: 1rem;
   }
-  
-  .label {
-    color:rgba(255, 255, 255, 0.8);
-    font-size: 20px;
-    font-weight: bold;
-  }
-  
-  .label-p{
-    display:inline-block;
-    color:rgba(255, 255, 255, 0.6);
-    font-size: 20px;
-  }
-  
-  .title{
-    color:white;
-    font-size: 28px;
-  }
+}
